scanner in progress

+++ tree_walk_interpreter/lox/src/error.rs
@@ -1,3 +1,3 @@
-pub fn report(line: i32, message: String, location: String) {
+pub fn report(line: usize, message: String, location: String) {
     println!("[line {}] Error {}: {}", line, location, message);
 }
